apps/speaking-python/requirements.txt
fastapi
uvicorn
openai
python-multipart
pydantic
python-dotenv